feat: Consolidate notification and logging systems
This commit introduces a unified, environment-aware logger and a centralized notification system to improve consistency and maintainability across the application.

Key changes include:

- A new `NotificationProvider` component to centralize the `notistack` configuration and provide a consistent look and feel for all notifications.
- A `useAppSnackbar` hook to simplify the notification API and address potential `exhaustive-deps` issues.
- A unified logger at `lib/logger.ts` that uses `pino` on the server and `console` on the client.
- Updated all application code, documentation, and test files to use the new logger.
- Added unit tests for the new logger to verify its behavior in both client and server environments.

Note: This commit includes breaking changes due to the removal of the old logger files. The build is currently failing due to an unresolved issue with conditionally importing the `pino` logging library in the project's ES module environment.
