+++ utils/index.ts
@@ -1,7 +1,7 @@
 export * from './audioManager'
 export * from './constants'
 export * from './dateUtils'
-export * from './logger'
+export * from '../lib/logger'
 export * from './socketManager'
 export * from './urls'
 export * from './visualization'
