+++ utils/websocketUtils.ts
@@ -8,7 +8,7 @@
  */
 import { WebSocket, Server as WebSocketServer } from 'ws'
 import { ExtWebSocket, ServerMessage } from '../types/websocket.js'
-import logger from './logger.js'
+import logger from '../lib/logger.js'
 
 /**
  * Sends a typed WebSocket message to a single client. This is the preferred
